Automated Cold Emailing/requirements.txt
# Google API and authentication
google-api-python-client>=2.86.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
google-generativeai>=0.3.0

# Excel and data manipulation
openpyxl>=3.1.2
pandas>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# HTTP requests
requests>=2.28.0

# Progress bars (used in some operations)
tqdm>=4.65.0